Feat: 모달 창 구현 및 Current Process 반영
