+++ src/components/ProcessSkipButton.tsx
@@ -1,6 +1,6 @@
 import { useMyPresence, useUpdateMyPresence } from "~/liveblocks.config";
 import useModalStore from "@/store/useModalStore";
-import Modal from "@/components/HelpModal";
+import Modal from "@/components/GuideModal";
 
 const ProcessSkipButton = () => {
   const [myPresence, ...rest] = useMyPresence();
